fix access denied error
in contrast to mysql_connect, it is not allowed to attach the port or the socket to the host name for the mysqli_connect command. There are separate parameters for that now. This leads to "Access denied for user 'user123'@'localhost' (using password: NO) " errors in the current version if port or socket is specified